+++ regcache.h
@@ -41,12 +41,12 @@ struct regcache;
 u8 lightrec_alloc_reg(struct regcache *cache, jit_state_t *_jit, u8 jit_reg);
 u8 lightrec_alloc_reg_temp(struct regcache *cache, jit_state_t *_jit);
 u8 lightrec_alloc_reg_out(struct regcache *cache, jit_state_t *_jit,
-			  u8 reg, u8 flags);
+			  u16 reg, u8 flags);
 u8 lightrec_alloc_reg_in(struct regcache *cache, jit_state_t *_jit,
-			 u8 reg, u8 flags);
+			 u16 reg, u8 flags);
 
 u8 lightrec_request_reg_in(struct regcache *cache, jit_state_t *_jit,
-			   u8 reg, u8 jit_reg);
+			   u16 reg, u8 jit_reg);
 
 s8 lightrec_get_reg_with_value(struct regcache *cache, intptr_t value);
 void lightrec_temp_set_value(struct regcache *cache, u8 jit_reg, intptr_t value);
@@ -66,11 +66,11 @@ void lightrec_storeback_regs(struct regcache *cache, jit_state_t *_jit);
 _Bool lightrec_has_dirty_regs(struct regcache *cache);
 
 void lightrec_clean_reg_if_loaded(struct regcache *cache, jit_state_t *_jit,
-				  u8 reg, _Bool unload);
-void lightrec_discard_reg_if_loaded(struct regcache *cache, u8 reg);
+				  u16 reg, _Bool unload);
+void lightrec_discard_reg_if_loaded(struct regcache *cache, u16 reg);
 
 u8 lightrec_alloc_reg_in_address(struct regcache *cache,
-		jit_state_t *_jit, u8 reg, s16 offset);
+		jit_state_t *_jit, u16 reg, s16 offset);
 
 struct native_register * lightrec_regcache_enter_branch(struct regcache *cache);
 void lightrec_regcache_leave_branch(struct regcache *cache,
